refactor: extract image menu component into separate file
